requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
absl-py==2.3.1
    # via ortools
alembic==1.16.5
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via uvicorn
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
deprecated==1.2.18
    # via nashpy
dill==0.4.0
    # via
    #   multiprocess
    #   pathos
fastapi==0.116.2
    # via -r requirements.in
filelock==3.19.1
    # via pref-voting
fonttools==4.60.0
    # via matplotlib
greenlet==3.2.4
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
immutabledict==4.2.1
    # via ortools
iniconfig==2.1.0
    # via pytest
kiwisolver==1.4.9
    # via matplotlib
llvmlite==0.44.0
    # via numba
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
matplotlib==3.10.6
    # via
    #   pref-voting
    #   seaborn
multiprocess==0.70.18
    # via pathos
nashpy==0.0.40
    # via pref-voting
networkx==3.5
    # via
    #   nashpy
    #   pref-voting
numba==0.61.2
    # via pref-voting
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   nashpy
    #   numba
    #   ortools
    #   pandas
    #   preflibtools
    #   prefsampling
    #   scipy
    #   seaborn
ortools==9.14.6206
    # via pref-voting
packaging==25.0
    # via
    #   matplotlib
    #   pytest
pandas==2.3.2
    # via
    #   -r requirements.in
    #   ortools
    #   seaborn
pathos==0.3.4
    # via pref-voting
pillow==11.3.0
    # via matplotlib
pluggy==1.6.0
    # via pytest
postmarker==1.0
    # via -r requirements.in
pox==0.3.6
    # via pathos
ppft==1.7.7
    # via pathos
pref-voting==1.16.31
    # via -r requirements.in
preflibtools==2.0.33
    # via pref-voting
prefsampling==0.1.24
    # via
    #   pref-voting
    #   preflibtools
protobuf==6.31.1
    # via ortools
psycopg2-binary==2.9.9
    # via -r requirements.in
pulp==3.3.0
    # via preflibtools
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pyparsing==3.2.4
    # via matplotlib
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==1.2.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via uvicorn
random2==1.0.2
    # via pref-voting
requests==2.32.5
    # via postmarker
scipy==1.16.2
    # via
    #   nashpy
    #   pref-voting
seaborn==0.13.2
    # via pref-voting
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements.in
    #   alembic
starlette==0.48.0
    # via fastapi
tabulate==0.9.0
    # via pref-voting
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   ortools
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wrapt==1.17.3
    # via deprecated
